# Backend/SmartInterviewAPI/DTOs/RegisterDto.cs
namespace SmartInterviewAPI.DTOs
{
    public class RegisterDto
    {
        public string FullName { get; set; }
        public string Email { get; set; }
        public string Password { get; set; }
        public string FrontendUrl { get; set; }

    }
}